Use Coro archive publishing in alto-coro
Pin alto-coro to the merged Coro archive/publish split and route soft block production through Coro-owned archived batches. Alto blocks are submitted as Coro transactions, grouped by Coro's batch policy, published as Celestia blobs, and recorded back into Coro as published cursors.

Update the explorer compatibility API and timing labels to reflect block-oriented Coro/Celestia semantics. Refresh the Mocha examples and READMEs so the fork is clearly presented as a single-writer Alto-shaped PoC, not a full Alto validator network.

Split the large alto-coro main file into focused modules for configuration, Alto block modeling, Celestia clients, HTTP history, the soft publish pipeline, and small utilities. This keeps the runtime flow unchanged while making the public code easier to read and review.
