+++ Cargo.toml
@@ -21,8 +21,8 @@ license = "MIT OR Apache-2.0"
 alto-chain = { version = "2026.5.0", path = "chain" }
 alto-client = { version = "2026.5.0", path = "client" }
 alto-types = { version = "2026.5.0", path = "types" }
-coro = { git = "ssh://git@github.com/celestiaorg/coro.git", rev = "665b7447710d4a1eaa12dc4e243592579746fe8d", package = "coro" }
-coro-demo = { git = "ssh://git@github.com/celestiaorg/coro.git", rev = "665b7447710d4a1eaa12dc4e243592579746fe8d", package = "coro-demo" }
+coro = { git = "ssh://git@github.com/celestiaorg/coro.git", rev = "2c195362884d146c8eea79cb16a2290237c7e4f8", package = "coro" }
+coro-demo = { git = "ssh://git@github.com/celestiaorg/coro.git", rev = "2c195362884d146c8eea79cb16a2290237c7e4f8", package = "coro-demo" }
 commonware-actor = "2026.5.0"
 commonware-broadcast = "2026.5.0"
 commonware-codec = "2026.5.0"
